--- a/crypto/bls12381_utils.c
+++ b/crypto/bls12381_utils.c
@@ -18,6 +18,10 @@ int get_invalid() {
     return INVALID;
 }
 
+void bn_new_wrapper(bn_t a) {
+    bn_new(a);
+}
+
 // global variable of the pre-computed data
 prec_st bls_prec_st;
 prec_st* bls_prec = NULL;
@@ -45,8 +49,8 @@ const uint64_t p_1div2_data[Fp_DIGITS] = {
 
 
 // sets the global variable to input
-void precomputed_data_set(prec_st* p) {
-    bls_prec = p;
+void precomputed_data_set(const prec_st* p) {
+    bls_prec = (prec_st*)p;
 }
 
 // Reads a prime field element from a digit vector in big endian format.
@@ -58,9 +62,11 @@ prec_st* init_precomputed_data_BLS12_381() {
     bls_prec = &bls_prec_st;
 
     #if (hashToPoint == OPSWU)
+
     fp_read_raw(bls_prec->a1, a1_data);
     fp_read_raw(bls_prec->b1, b1_data);
     // (p-3)/4
+    bn_new(&bls_prec->p_3div4);
     bn_read_raw(&bls_prec->p_3div4, p_3div4_data, Fp_DIGITS);
     // (p-1)/2
     fp_read_raw(bls_prec->fp_p_1div2, fp_p_1div2_data);
@@ -75,10 +81,13 @@ prec_st* init_precomputed_data_BLS12_381() {
     #endif
 
     #if (MEMBERSHIP_CHECK_G1 == BOWE)
+    bn_new(&bls_prec->beta);
     bn_read_raw(&bls_prec->beta, beta_data, Fp_DIGITS);
+    bn_new(&bls_prec->z2_1_by3);
     bn_read_raw(&bls_prec->z2_1_by3, z2_1_by3_data, 2);
     #endif
 
+    bn_new(&bls_prec->p_1div2);
     bn_read_raw(&bls_prec->p_1div2, p_1div2_data, Fp_DIGITS);
     return bls_prec;
 }
@@ -222,7 +231,7 @@ void bn_map_to_Zr_star(bn_t a, const uint8_t* bin, int len) {
 
 // returns the sign of y.
 // 1 if y > (p - 1)/2 and 0 otherwise.
-static int fp_get_sign(fp_t y) {
+static int fp_get_sign(const fp_t y) {
     bn_t bn_y;
     bn_new(bn_y);
     fp_prime_back(bn_y, y);