--- a/crypto/build_dependency.sh
+++ b/crypto/build_dependency.sh
@@ -3,7 +3,7 @@
 set -euo pipefail
 
 # relic version or tag
-relic_version="7a9bba7f"
+relic_version="9206ae50"
 
 rm -rf relic