The diff contains:
- a lot of changes that concern parts of the library we don't use (other curves a la BN, BLS24-X, BLS12-383 ...), integer protocols (ETRS), field extension machinery ...
- otherwise irrelevant changes, e.g. CI/CD
- some memory bug fixing

[Formatted Log](https://gist.githubusercontent.com/huitseeker/e8bbd3f0475ebe93277bdfd46d9424de/raw/385a3dba9fc927b76f1c04e4e96474bff99cab7c/gistfile1.txt)
[Full Changeset](https://github.com/relic-toolkit/relic/compare/7a9bba7f..9206ae5)

**Benchmark Diff:**

`benchstat ~/old_relic.txt ~/new_relic.txt `

```
name                           old time/op  new time/op  delta
ScalarMult/G1-8                 144µs ± 0%   142µs ± 0%   ~     (p=1.000 n=1+1)
ScalarMult/G2-8                 329µs ± 0%   326µs ± 0%   ~     (p=1.000 n=1+1)
HashToG1-8                      231µs ± 0%   230µs ± 0%   ~     (p=1.000 n=1+1)
CheckG1-8                       210µs ± 0%   209µs ± 0%   ~     (p=1.000 n=1+1)
BLSBLS12381Sign-8               477µs ± 0%   473µs ± 0%   ~     (p=1.000 n=1+1)
BLSBLS12381Verify-8            2.71ms ± 0%  2.64ms ± 0%   ~     (p=1.000 n=1+1)
BatchVerify/happy_path-8       81.5ms ± 0%  81.5ms ± 0%   ~     (p=1.000 n=1+1)
BatchVerify/unhappy_path-8      117ms ± 0%   115ms ± 0%   ~     (p=1.000 n=1+1)
VerifySignatureManyMessages-8  3.14ms ± 0%  3.12ms ± 0%   ~     (p=1.000 n=1+1)
Aggregate/PrivateKeys-8        71.6µs ± 0%  70.1µs ± 0%   ~     (p=1.000 n=1+1)
Aggregate/PublicKeys-8         2.85ms ± 0%  2.81ms ± 0%   ~     (p=1.000 n=1+1)
Aggregate/Signatures-8         47.9ms ± 0%  48.0ms ± 0%   ~     (p=1.000 n=1+1)
SimpleKeyGen-8                 20.2ms ± 0%  20.3ms ± 0%   ~     (p=1.000 n=1+1)
```

**Fixed bugs:**

- Unexpected failure of ep2\_mul\[\_lwnaf\] above the prime group order [\#64](https://github.com/relic-toolkit/relic/issues/64)

**Closed issues:**

- Other way to construct towered extension fields [\#203](https://github.com/relic-toolkit/relic/issues/203)
- blake2.h:101:5: error: size of array element is not a multiple of its alignment [\#202](https://github.com/relic-toolkit/relic/issues/202)
- ECIES 160bit [\#201](https://github.com/relic-toolkit/relic/issues/201)
- Compilation with "ARITH gmp" fails [\#200](https://github.com/relic-toolkit/relic/issues/200)
- Support for armv8-a ? [\#198](https://github.com/relic-toolkit/relic/issues/198)
- Function name bn\_init conflicts with OpenSSL when used in tandem [\#196](https://github.com/relic-toolkit/relic/issues/196)
- 16-bit MSP430 [\#193](https://github.com/relic-toolkit/relic/issues/193)
- Modular exponentiation returns 1 if exponent is 0 and modulo is 1 [\#185](https://github.com/relic-toolkit/relic/issues/185)
- Compilation of RELIC with bls12-446 and bls12-455 fails [\#182](https://github.com/relic-toolkit/relic/issues/182)
- test\_bn fails with BLS12-381 preset [\#181](https://github.com/relic-toolkit/relic/issues/181)
- \[BUG\] undefined reference to `bench_init', `bench\_clean' [\#180](https://github.com/relic-toolkit/relic/issues/180)
- Tests FTBFS because of missing symbol in header [\#179](https://github.com/relic-toolkit/relic/issues/179)
- Builds are broken [\#178](https://github.com/relic-toolkit/relic/issues/178)
- compile error  inlining failed in call to always\_inline ‘\_mm\_alignr\_epi8’ on unbantu20.04 gcc9 [\#177](https://github.com/relic-toolkit/relic/issues/177)
- bn\_write\_str buffer overflow [\#176](https://github.com/relic-toolkit/relic/issues/176)
- ECDSA verify succeeds when it should fail [\#175](https://github.com/relic-toolkit/relic/issues/175)
- ec\_mul\_gen hangs with curve SECG\_K256 [\#174](https://github.com/relic-toolkit/relic/issues/174)
- Wrong square root computation [\#173](https://github.com/relic-toolkit/relic/issues/173)
- Out-of-bounds read via bn\_sqr\_basic [\#172](https://github.com/relic-toolkit/relic/issues/172)
- OSS-Fuzz integration [\#171](https://github.com/relic-toolkit/relic/issues/171)
- Building Relic with Curve NIST\_P256 throws FATAL ERROR in relic\_fp\_prime.c:120 [\#170](https://github.com/relic-toolkit/relic/issues/170)
- Compressing \(packing\) a point to binary array does not comply with X9.62 standard [\#169](https://github.com/relic-toolkit/relic/issues/169)
-  ‘ctx\_t’ {aka ‘struct \_ctx\_t’} has no member named ‘total’ [\#168](https://github.com/relic-toolkit/relic/issues/168)
- relic does not work with C++ [\#167](https://github.com/relic-toolkit/relic/issues/167)
- Memory leak in ep2\_curve\_init/clean with ALLOC=DYNAMIC [\#166](https://github.com/relic-toolkit/relic/issues/166)
- \*\_is\_valid\(\) functions produce false negative for not normalized points [\#147](https://github.com/relic-toolkit/relic/issues/147)
- Bench and Test doesnt build [\#122](https://github.com/relic-toolkit/relic/issues/122)

**Merged pull requests:**

- Add pairing delegation protocols [\#199](https://github.com/relic-toolkit/relic/pull/199) ([dfaranha](https://github.com/dfaranha))
- Fix support for Win64/MSVC targets. [\#197](https://github.com/relic-toolkit/relic/pull/197) ([dfaranha](https://github.com/dfaranha))
- Simplify generator getting for Gt. [\#194](https://github.com/relic-toolkit/relic/pull/194) ([luozejiaqun](https://github.com/luozejiaqun))
- cmake: Always use user defined CFLAGS, not only for release builds [\#187](https://github.com/relic-toolkit/relic/pull/187) ([xdustinface](https://github.com/xdustinface))
- Fix MinGW build [\#186](https://github.com/relic-toolkit/relic/pull/186) ([xdustinface](https://github.com/xdustinface))
- Remove debug printf in bn\_mxp\_slide [\#184](https://github.com/relic-toolkit/relic/pull/184) ([guidovranken](https://github.com/guidovranken))
- Remove ALLOC = STACK to simplify memory allocation. [\#183](https://github.com/relic-toolkit/relic/pull/183) ([dfaranha](https://github.com/dfaranha))
- Update relic\_alloc.h [\#165](https://github.com/relic-toolkit/relic/pull/165) ([aguycalled](https://github.com/aguycalled))
- Add correct support for FreeBSD and NetBSD [\#164](https://github.com/relic-toolkit/relic/pull/164) ([hoffmang9](https://github.com/hoffmang9))