@huitseeker 2021-06-16
@tarakby I'm looking at the failing test, which is a change in the semantics of interpreting an empty slice as zero since https://github.com/relic-toolkit/relic/commit/16f3c55b055117e1e4cc82160675e2a778a2ed6e (see the last file)