@tarakby 2021-07-12
Thanks @huitseeker for adding the commits!

Everything looks good, we just need to delete the 2 lines commented [here](https://github.com/onflow/flow-go/pull/846#discussion_r667263222) and [here](https://github.com/onflow/flow-go/pull/846#discussion_r667263592) since the initializations are now happening before calling the functions.